+++ src/pyprefab/cli.py
@@ -9,10 +9,15 @@ import typer
 from jinja2 import Environment, FileSystemLoader
 from rich import print
 from rich.panel import Panel
+from typing_extensions import Annotated
 
 logger = structlog.get_logger()
 
-app = typer.Typer(help='Generate python project scaffolding based on pyprefab.')
+app = typer.Typer(
+    add_completion=False,
+    help='Generate python project scaffolding based on pyprefab.',
+    rich_markup_mode='markdown',
+)
 
 
 def validate_project_name(name: str) -> bool:
@@ -60,20 +65,37 @@ def render_templates(context: dict, templates_dir: Path, target_dir: Path):
 
 @app.command()
 def create(
-    name: str = typer.Argument(..., help='Name of the project'),
-    author: str = typer.Option(..., '--author', help='Project author'),
-    description: str = typer.Option('', '--description', help='Project description'),
-    project_dir: Optional[Path] = typer.Option(
-        None, '--directory', help='Directory that will contain the project (defaults to current directory)'
-    ),
+    name: Annotated[str, typer.Argument(help='Name of the project', show_default=False)],
+    author: Annotated[Optional[str], typer.Option(help='Project author', prompt='Project author', show_default=False)],
+    description: Annotated[
+        Optional[str], typer.Option(help='Project description', prompt='Project description', show_default=False)
+    ],
+    project_dir: Annotated[
+        Path,
+        typer.Option(
+            '--dir',
+            help='Directory that will contain the project',
+            prompt='Project directory',
+            show_default=False,
+        ),
+    ],
 ):
-    """Generate a new Python project from templates."""
+    """
+    :snake: **Create Python package boilerplate** :snake:
+    """
     if not validate_project_name(name):
         typer.secho(
             f'Error: {name} is not a valid Python package name',
             fg=typer.colors.RED,
         )
         raise typer.Exit(1)
+
+    if project_dir.exists() and any(project_dir.iterdir()):
+        proceed = typer.confirm(f'Directory {project_dir} is not empty. Proceed?')
+        if not proceed:
+            typer.secho(f'{name} not created', fg=typer.colors.YELLOW)
+            raise typer.Exit(code=1)
+
     templates_dir = Path(__file__).parent / 'templates'
     target_dir = project_dir or Path.cwd() / name