+++ CHANGELOG.md
@@ -4,12 +4,25 @@ All notable changes to `pyprefab` are documented here.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and the project uses [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.4.0] - [2025-01-26]
 
 ### Added
 
 - New templates for GitHub workflows: ci (linting, tests, and coverage), publishing to TestPyPI, and publishing to PyPI
 - Project changelog
+- New CHANGELOG.md template
+- New CONTRIBUTING.md template
+
+### Changed
+
+- Add CLI prompts as an alternate to passing command options
+- `--directory` option renamed to `--dir`
+- Prompt user when specified project directory is not empty
+- List created package components on README.md
+
+### Fixed
+
+- Correct CLI example on README.md
 
 ## [0.3.2] - [2025-01-13](https://github.com/bsweger/pyprefab/compare/v0.3.1...v0.3.2)